+++ remote/remote_http_test.go
@@ -15,6 +15,20 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+func Test_newHTTPRemote(t *testing.T) {
+	t.Run("should return an error if creating the backoff fails", func(t *testing.T) {
+		config := &core.Remote{
+			RetryPolicy: core.RetryPolicy{
+				Interval: -2,
+			},
+		}
+
+		_, err := newHTTPRemote(config, nil)
+
+		require.Error(t, err)
+	})
+}
+
 func Test_checkStatusCode(t *testing.T) {
 	t.Run("should return nil for HTTP 200", func(t *testing.T) {
 		mockResp := &http.Response{}